+++ Hanabi/src/Engine/Renderer/Renderer2D.h
@@ -51,9 +51,9 @@ namespace Hanabi
 		static void ResetStats();
 		static Statistics GetStats();
 	private:
-		static float GetTextureID(const Ref<Texture2D>& texture);
+		static uint32_t GetTextureID(const Ref<Texture2D>& texture);
 		static void SetQuadVertex(const glm::mat4& transform, const glm::vec4& color,
-			const glm::vec2* texCoord, float texIndex, float tilingFactor);
+			const glm::vec2* texCoord, uint32_t texIndex, float tilingFactor);
 		static void StartBatch();
 		static void Flush();
 		static void NextBatch();
